pyproject.toml
[build-system]
requires = ["google", "setuptools>=60", "setuptools_scm[toml]>=7", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memorisdk"
authors = [{ name="Memori", email="noc@gibsonai.com" }]
description = "Memori Python SDK"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Operating System :: OS Independent",
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.9",
]
version = "3.0.0"
dependencies = [
    "aiohttp",
    "botocore",
    "grpcio",
    "protobuf",
    "pyfiglet"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
memori = ["data/*"]

[project.urls]
Homepage = "https://memorilabs.ai/"
Documentation = "https://developers.memorilabs.ai/"
Repository = "https://github.com/GibsonAI/"

[tool.black]
line-length=88
target-version = ['py312']

[tool.isort]
profile = "black"
